requirements.txt
Django==2.2
djangorestframework==3.9.2
psycopg2==2.7.6.1
pytz==2019.1
sqlparse==0.3.0
